+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 ## PHP xdebug (dbgp) proxy ChangeLog
 
-### master
+### [Unreleased]
 
 There are next changes:
 
-### v0.4.0
+### [v0.4.1]
+
+There are next changes:
+
+- Fix changes in grumphp and bad composer.lock ([#26](https://github.com/mougrim/php-xdebug-proxy/pull/26))
+
+### [v0.4.0]
 
 There are next changes:
 
@@ -16,7 +22,7 @@ There are next changes:
 - parameter $config was added to method `\Mougrim\XdebugProxy\Factory\Factory::createRequestPreparers()` ([#22](https://github.com/mougrim/php-xdebug-proxy/pull/22))
 - now `\Mougrim\XdebugProxy\Factory\SoftMocksFactory::createConfig()` should return `\Mougrim\XdebugProxy\Config\SoftMocksConfig` ([#22](https://github.com/mougrim/php-xdebug-proxy/pull/22))
 
-### v0.3.0
+### [v0.3.0]
 
 There are next changes:
 
@@ -27,13 +33,13 @@ There are next changes:
 - constants visibility was added ([#18](https://github.com/mougrim/php-xdebug-proxy/pull/18))
 - deprecated interface \Mougrim\XdebugProxy\RequestPreparer was removed ([#18](https://github.com/mougrim/php-xdebug-proxy/pull/18))
 
-### v0.2.1
+### [v0.2.1]
 
 There are next changes:
 
 - possibility to disable IDE registration was added ([#12](https://github.com/mougrim/php-xdebug-proxy/pull/12))
 
-### v0.2.0
+### [v0.2.0]
 
 There are next changes:
 
@@ -44,7 +50,7 @@ There are next changes:
 - [README.md](README.md) was updated ([#11](https://github.com/mougrim/php-xdebug-proxy/pull/11))
 - doc for [`RequestPreparer\RequestPreparer`](src/RequestPreparer/RequestPreparer.php) was added ([#11](https://github.com/mougrim/php-xdebug-proxy/pull/11))
 
-### v0.1.0
+### [v0.1.0]
 
 There are next changes:
 
@@ -53,3 +59,11 @@ There are next changes:
 - now `\Mougrim\XdebugProxy\RequestPreparer\RequestPreparer` is used instead of `\Mougrim\XdebugProxy\RequestPreparer` ([#7](https://github.com/mougrim/php-xdebug-proxy/pull/7))
 - `\Mougrim\XdebugProxy\RequestPreparer\RequestPreparer` methods can throw `\Mougrim\XdebugProxy\RequestPreparer\Error` and `\Mougrim\XdebugProxy\RequestPreparer\Exception` if there is some problem ([#7](https://github.com/mougrim/php-xdebug-proxy/pull/7))
 - now `\Mougrim\XdebugProxy\Factory\Factory::createRequestPreparers()` accepts `$logger` param and also can throw `\Mougrim\XdebugProxy\RequestPreparer\Error` and `\Mougrim\XdebugProxy\RequestPreparer\Exception` if there is some problem ([#7](https://github.com/mougrim/php-xdebug-proxy/pull/7))
+
+[unreleased]: https://github.com/mougrim/php-xdebug-proxy/compare/0.4.1...HEAD
+[v0.4.1]: https://github.com/mougrim/php-xdebug-proxy/compare/0.4.0...0.4.1
+[v0.4.0]: https://github.com/mougrim/php-xdebug-proxy/compare/0.3.0...0.4.0
+[v0.3.0]: https://github.com/mougrim/php-xdebug-proxy/compare/0.2.1...0.3.0
+[v0.2.1]: https://github.com/mougrim/php-xdebug-proxy/compare/0.2.0...0.2.1
+[v0.2.0]: https://github.com/mougrim/php-xdebug-proxy/compare/0.1.0...0.2.0
+[v0.1.0]: https://github.com/mougrim/php-xdebug-proxy/compare/0.0.1...0.1.0